Create An overview of html API
You will be getting an overview of what html API is. Also by the time you must have complete this article you will be acquainted with the types of html API, it purpose, usefulness and applications. 

Html API will definitely mean the that allows you to freely use the special  features of html technology with ease. 

API is application programming interface , that a type of technology that allows other software application to interact or extract data from other applications of the internet. 
Html have